testing with simplejson

+++ lambda/put-game/coordinates.py
@@ -1,3 +1,6 @@
+from decimal import Decimal
+
+
 def from_numbers_to_strings(game):
     game['startPos']['long'] = str(game['startPos']['long'])
     game['startPos']['lat'] = str(game['startPos']['lat'])
@@ -9,10 +12,10 @@ def from_numbers_to_strings(game):
 
 
 def from_strings_to_numbers(game):
-    game['startPos']['long'] = float(game['startPos']['long'])
-    game['startPos']['lat'] = float(game['startPos']['lat'])
+    game['startPos']['long'] = Decimal(game['startPos']['long'])
+    game['startPos']['lat'] = Decimal(game['startPos']['lat'])
     for path in game['paths']:
         for point in path:
-            point['long'] = float(point['long'])
-            point['lat'] = float(point['lat'])
+            point['long'] = Decimal(point['long'])
+            point['lat'] = Decimal(point['lat'])
     return game
